fix: refactor package structure
