+++ graphql-relay/controllers/user.resolvers.go
@@ -12,7 +12,6 @@ import (
 	"github.com/Reimei1213/lab/graphql-relay/interface/presenter"
 	"github.com/Reimei1213/lab/graphql-relay/pkg/graph"
 	"github.com/Reimei1213/lab/graphql-relay/pkg/graph/model"
-	"github.com/Reimei1213/lab/graphql-relay/pkg/graph/pagination"
 )
 
 // AddUser is the resolver for the addUser field.
@@ -22,7 +21,7 @@ func (r *mutationResolver) AddUser(ctx context.Context, input model.AddUserInput
 		slog.Error("failed to create user", "err", err)
 		return nil, err
 	}
-	return &model.AddUserPayload{ID: pagination.EncodeGraphqlID(presenter.NodeTypeUser, id)}, nil
+	return &model.AddUserPayload{ID: graph.EncodeGraphqlID(presenter.NodeTypeUser, id)}, nil
 }
 
 // Users is the resolver for the users field.
@@ -30,14 +29,14 @@ func (r *queryResolver) Users(ctx context.Context, first *int, after *string, la
 	var afterCursor, beforeCursor *string
 
 	if after != nil {
-		_, id, err := pagination.DecodeGraphqlID(*after)
+		_, id, err := graph.DecodeGraphqlID(*after)
 		if err != nil {
 			return nil, err
 		}
 		afterCursor = &id
 	}
 	if before != nil {
-		_, id, err := pagination.DecodeGraphqlID(*before)
+		_, id, err := graph.DecodeGraphqlID(*before)
 		if err != nil {
 			return nil, err
 		}
@@ -48,7 +47,7 @@ func (r *queryResolver) Users(ctx context.Context, first *int, after *string, la
 		slog.Error("failed to get users", "err", err)
 		return nil, err
 	}
-	return pagination.NewConnection(presenter.ToUsers(res.Users), res.HasNextPage, res.HasPreviousPage), nil
+	return presenter.NewConnection(presenter.ToUsers(res.Users), res.HasNextPage, res.HasPreviousPage), nil
 }
 
 // Cards is the resolver for the cards field.
