+++ graphql-relay/interface/presenter/pagination.go
@@ -1,37 +1,16 @@
-package pagination
+package presenter
 
 import (
-	"encoding/base64"
-	"fmt"
-	"strings"
-
+	"github.com/Reimei1213/lab/graphql-relay/pkg/graph"
 	"github.com/Reimei1213/lab/graphql-relay/pkg/graph/model"
 )
 
-type NodeType string
+type NodeType = graph.NodeType
 
 type Node interface {
 	GetID() string
 	GetNodeType() NodeType
 	ToGraphqlNode() model.Node
-}
-
-const graphqlIDFormat = "%s:%s" // tableID:id
-
-func EncodeGraphqlID(nodeType NodeType, id string) string {
-	graphqlID := fmt.Sprintf(graphqlIDFormat, nodeType, id)
-	return base64.StdEncoding.EncodeToString([]byte(graphqlID))
-}
-
-func DecodeGraphqlID(encodedGraphqlID string) (NodeType, string, error) {
-	graphqlID, err := base64.StdEncoding.DecodeString(encodedGraphqlID)
-	if err != nil {
-		return "", "", err
-	}
-
-	result := strings.Split(string(graphqlID), ":")
-
-	return NodeType(result[0]), result[1], nil
 }
 
 func toGraphqlModels[T Node](nodes []T) []model.Node {
@@ -54,7 +33,7 @@ func newEdges[T Node](nodes []T) []*model.Edge {
 	res := make([]*model.Edge, 0, len(nodes))
 	for _, n := range nodes {
 		res = append(res, &model.Edge{
-			Cursor: EncodeGraphqlID(n.GetNodeType(), n.GetID()),
+			Cursor: graph.EncodeGraphqlID(n.GetNodeType(), n.GetID()),
 			Node:   n.ToGraphqlNode(),
 		})
 	}
@@ -71,8 +50,8 @@ func newPageInfo[T Node](nodes []T, hasNextPage, hasPreviousPage bool) *model.Pa
 
 	startNode := nodes[0]
 	endNode := nodes[len(nodes)-1]
-	startCursor := EncodeGraphqlID(startNode.GetNodeType(), startNode.GetID())
-	endCursor := EncodeGraphqlID(endNode.GetNodeType(), endNode.GetID())
+	startCursor := graph.EncodeGraphqlID(startNode.GetNodeType(), startNode.GetID())
+	endCursor := graph.EncodeGraphqlID(endNode.GetNodeType(), endNode.GetID())
 	return &model.PageInfo{
 		HasPreviousPage: hasPreviousPage,
 		StartCursor:     &startCursor,
