+++ graphql-relay/interface/presenter/user.go
@@ -2,23 +2,23 @@ package presenter
 
 import (
 	"github.com/Reimei1213/lab/graphql-relay/domain/entity"
+	"github.com/Reimei1213/lab/graphql-relay/pkg/graph"
 	"github.com/Reimei1213/lab/graphql-relay/pkg/graph/model"
-	"github.com/Reimei1213/lab/graphql-relay/pkg/graph/pagination"
 )
 
-const NodeTypeUser pagination.NodeType = "users"
+const NodeTypeUser NodeType = "users"
 
 type User struct {
 	*entity.User
 }
 
-var _ pagination.Node = (*User)(nil)
+var _ Node = (*User)(nil)
 
 func (u *User) GetID() string {
 	return u.ID
 }
 
-func (u *User) GetNodeType() pagination.NodeType {
+func (u *User) GetNodeType() NodeType {
 	return NodeTypeUser
 }
 
@@ -27,7 +27,7 @@ func (u *User) ToGraphqlModel() *model.User {
 		return nil
 	}
 	return &model.User{
-		ID:   pagination.EncodeGraphqlID(u.GetNodeType(), u.ID),
+		ID:   graph.EncodeGraphqlID(u.GetNodeType(), u.ID),
 		Name: u.Name,
 	}
 }
